adding the decoding to the existing process_instruction in processor.rs along with slot_hashes.bin in the same dir
